requirements.txt
sqlalchemy
psycopg2-binary
numpy
falcon
prefect
streamlit
python-dotenv